added handlers to homepage
