+++ Frontend/src/app/listings/util.js
@@ -1,0 +1,35 @@
+/*
+  function to handle messaging seller, used when Message Seller button on a card is clicked
+  Still needs to be defined with proper parameters and backend endpoint
+  */
+const handleMessageSeller = async () => {
+  try {
+    // console.log("messaging seller"); // debug log
+    const res = await fetch("http://localhost:8080/", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify(),
+    });
+  } catch (err) {
+    console.error("Error messaging seller:", err);
+  }
+};
+
+/*
+  function to handle booking a listing, used when the book button on a card is clicked
+  Still needs to be defined with proper parameters and backend endpoint
+  */
+const handleBook = async () => {
+  // console.log("booking listing"); // debug log
+  try {
+    const res = await fetch("http://localhost:8080/", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify(),
+    });
+  } catch (err) {
+    console.error("Error creating booking:", err);
+  }
+};
+
+export { handleMessageSeller, handleBook };
